Intermediate class for sales document
